# packages/openapi/openapi.json
{
  "openapi": "3.0.2",
  "paths": {
    "/api/v1/auth/me": {
      "get": {
        "description": "Get the currently authenticated user from the JWT bearer token",
        "summary": "Get authenticated user",
        "tags": [
          "Auth"
        ],
        "parameters": [],
        "operationId": "getMe",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": {
                      "type": "string"
                    },
                    "user_role": {
                      "type": "string"
                    },
                    "permissions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "user_id",
                    "permissions"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/auth/oauth/login": {
      "get": {
        "description": "Start the OAuth authorization code flow. By default this redirects to the provider. Use response=json to get the authorization URL in OpenAPI.",
        "summary": "Start OAuth login",
        "tags": [
          "Auth"
        ],
        "parameters": [],
        "operationId": "startOAuth",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "auth_url": {
                      "type": "string",
                      "format": "uri"
                    },
                    "state": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "auth_url",
                    "state"
                  ]
                }
              }
            }
          },
          "302": {
            "description": "302",
            "content": {
              "application/json": {
                "schema": {
                  "nullable": true
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/oauth/callback": {
      "get": {
        "description": "Complete the OAuth authorization code flow by exchanging code for provider tokens.",
        "summary": "Complete OAuth callback",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "operationId": "completeOAuth",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string"
                    },
                    "token_type": {
                      "type": "string"
                    },
                    "expires_in": {
                      "type": "number"
                    },
                    "refresh_token": {
                      "type": "string"
                    },
                    "scope": {
                      "type": "string"
                    },
                    "id_token": {
                      "type": "string"
                    },
                    "raw": {
                      "nullable": true
                    }
                  },
                  "required": [
                    "access_token",
                    "token_type"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/service-token/status": {
      "get": {
        "description": "Validate an opaque service token sent in the x-service-token header.",
        "summary": "Validate service token",
        "tags": [
          "Auth"
        ],
        "parameters": [],
        "operationId": "getServiceTokenStatus",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "authenticated": {
                      "type": "boolean"
                    },
                    "auth_type": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "authenticated",
                    "auth_type",
                    "message"
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "x-service-token": []
          }
        ]
      }
    },
    "/status": {
      "get": {
        "description": "Get health status",
        "summary": "Get health",
        "tags": [
          "Health"
        ],
        "parameters": [],
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "healthy",
                        "unhealthy"
                      ]
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "environment": {
                      "type": "string"
                    },
                    "checks": {
                      "type": "object",
                      "properties": {
                        "database": {
                          "type": "object",
                          "properties": {
                            "status": {
                              "type": "string"
                            },
                            "response_time": {
                              "type": "string"
                            },
                            "error": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "status",
                            "response_time"
                          ]
                        },
                        "redis": {
                          "type": "object",
                          "properties": {
                            "status": {
                              "type": "string"
                            },
                            "response_time": {
                              "type": "string"
                            },
                            "error": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "status",
                            "response_time"
                          ]
                        }
                      },
                      "required": [
                        "database"
                      ]
                    }
                  },
                  "required": [
                    "status",
                    "timestamp",
                    "environment",
                    "checks"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "info": {
    "version": "1.0.0",
    "title": "Boilerplate REST API - Documentation",
    "description": "Boilerplate REST API - Documentation"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "x-service-token": {
        "type": "apiKey",
        "name": "x-service-token",
        "in": "header"
      }
    }
  }
}